// src/app/model/ipropertyBase.ts
export interface IpropertyBase{
        id: number;
        sellRent: number;
        name: string;
        pType: string;
        fType: string;
        price: number;
        BHK: number;
        builtArea: number;
        city: string;
        RTM: number;
        image?: string;
      }  //it is an optional Field so that's y keep ? question mark